requirements.txt
numpy>=1.26,<3
opencv-python>=4.10,<5
fastapi>=0.115,<1
uvicorn>=0.30,<1
pydantic>=2.8,<3
Pillow>=10.4,<11
ultralytics>=8.3,<9
segment-anything>=1.0
streamlit>=1.39,<2
python-multipart>=0.0.9,<1
